Split VersionMismatchFinder.ts into separate files.
